test: move functional write concern error tests our of unit tests
